user quarantine: use raw pmail for ticket assembly
Currently, the quarantine report does not work if the recipient has
some encodable characters in their local part - e.g.
'some&other@domain.example'

When clicking on the links on the report the user gets still logged
in, the ticket _is_ valid after all, however their quarantine list is
empty, as the API call to `/quarantine/spamusers` returns 403 due to
the (encoded) username from the ticket not matching the (by the API
decoded) one from the request quarantine.

With this patch the username, which is includes in the ticket,
remains 'some&other@domain.example' instead of the encoded
'some&amp;other@domain.example', thus the access check user
comparission work with the correct value again and the listing works
as expected

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
 [ TL: commit message additions and rewordings ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>