fix #2606: ruledb disclaimer: add ability to set position
there are usecases, where one wants to embed a disclaimer (or similar)
text or html at the beginning of the mail content, not at the end.

so introduce a 'position' parameter for the disclaimer that does exactly
that.

internally this will get saved as `top` only if it is not the default of '0'.
We save it in the `Attribut` table in that case.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>