filter: skip warning about max_filters on low-memory if manually set
The get_max_filters method gets called on module load to fill the
default value used, at that stage we cannot know if the admin set
max_filters manually or not, so the warning might be bogus and thus
annoying.

Move it over to the smtp-filter executable's startup code, as there we
can check the config if it's max_filters is set or not.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>