d/triggers: add interest in spamassassin updates
currently an update to SpamAssassin does not reload any of the
services.

This patch simply adds interest-noawait on a trigger activated by
proxmox-spamassassin, which in turn reloads all services (as already
done for updates to pve-common, and in PVE on updates on other
packages that affect the API).

Technically it might only be necessary to reload pmg-smtp-filter,
pmgproxy and pmgdaemon, to get the benefits of a new spamassassin
version, but reloading all services seems more robust in the long run.

[0] https://git.dpkg.org/cgit/dpkg/dpkg.git/tree/doc/spec/triggers.txt
[1] https://manpages.debian.org/bookworm/dpkg-dev/deb-triggers.5.en.html

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>