api: remove immediate download path returns
always return an object with a download member instead, like we do in PVE.
without this cleanup (and corresponding removal of the special handling in
PMG's HttpServer), endpoints that can either download or return data like the
task log one cannot accomodate the new stricter download endpoint checks in
pve-http-server.

since the download information is directly consumed by our HTTP stack and not
actually returned to the HTTP client, this is not a breaking change of the
return value, but just affects the internal implementation.

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by:  Dominik Csapak <d.csapak@proxmox.com>