ruledb remove: honor content-disposition: 'attachment'
While this should seldomly happen with regular mail, having an
attachment (content-disposition header set to attachment) as first or
only part is possible, and users reasonably expect the part to be
removed if the action has 'remove all attachments' set.

See https://metacpan.org/pod/MIME::Entity and
https://datatracker.ietf.org/doc/html/rfc2183.html
for further information on content-disposition.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>