ruledb remove: treat single-part attachment as attachment
ignore the special case of treating skipping single-part messages in a
remove-matching action, if they have a text/* content-type, if the
mail explicitly says the part is an attachment.

seen with certain spam/phishing mails, which probably try to use this
as obfuscation technique.

additionally add a comment that MIME::Head::mime_type falls back to
text/plain if no content-type header is present (in accordance with
https://datatracker.ietf.org/doc/html/rfc2046#section-5.1), as I have
looked this up at least 5 times over the past years.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>