pmg7to8: add check for deprecated default entries in ruledb
We recently switched the example rule to use a domain that won't ever
be valid, but that only covers new setups. For existing setups we add
a check to our upgrade helper script to detect such entries.
Put that test in a new section for future checks of the rule database.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>