ruledb: content-type: do not rely on PMX_header_ct
As noted in the cover-letter for v3 of the series introducing this [0]
the PMX_magic_ct, PMX_glob_ct and PMX_header_ct are only set if there
is a bodyhandle for a mime-part, which only exists if the part has no
sub-parts.

[0]: https://lore.proxmox.com/pmg-devel/74a089b6-d858-4525-9822-da58b9c8f4d7@proxmox.com/T/#t

However we want to match on content-types even for mail-parts which
have no bodyhandle:
* our internal regression-test suite has a mail with `message/partial`
  set
* there are probably sensible uses for matching `multipart/related`
  parts (e.g. when trying to get a negated match.

Noticed during a check of our regression-tests.

Fixes: c320d56 ("utils: content-type: don't fallback to header information for magic")
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Reviewed-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>